Added ContentHandler
ContentHandler is an easy to extend, versatile handler of input files. Currently it is only used to load, store and supply images, but it could be used for any type of content.

Updated all other classes to make use of ContentHandler, removed the global variables that ContentHandler has replaced.

Added an Object class for gameobjects to inherit, reducing duplicate code

Moved the animation image array from Player to GameState because it was being troublesome

Removed useless Score class